+++ modules/aws-asg/versions.tf
@@ -1,12 +1,10 @@
 terraform {
-  required_version = "~> 0.13"
+  required_version = "~> 0.14"
 
   required_providers {
-    aws      = "~> 2"
+    aws      = "~> 3.26"
+    null     = "~> 3"
+    random   = "~> 3"
     template = "~> 2"
   }
-}
-
-provider "aws" {
-  region = var.aws_region
 }